Narrow TID16/SID16 to ushort, remove casts

+++ PKHeX.Core/Editing/PKM/EntitySummary.cs
@@ -81,8 +81,8 @@ public class EntitySummary // do NOT seal, allow inheritance
     public bool IsNicknamed => pk.IsNicknamed;
     public bool IsShiny => pk.IsShiny;
 
-    public uint TID16 => pk.TID16;
-    public uint SID16 => pk.SID16;
+    public ushort TID16 => pk.TID16;
+    public ushort SID16 => pk.SID16;
     public uint TSV => pk.TSV;
     public int Move1_PP => pk.Move1_PP;
     public int Move2_PP => pk.Move2_PP;
