+++ PKHeX.Core/Legality/Encounters/EncounterStatic/EncounterMight9.cs
@@ -284,6 +284,6 @@ public sealed record EncounterMight9 : EncounterStatic, ITeraRaid9
         0 => PersonalInfo.RatioMagicMale,
         1 => PersonalInfo.RatioMagicFemale,
         2 => PersonalInfo.RatioMagicGenderless,
-        _ => (byte)PersonalTable.SV.GetFormEntry(Species, Form).Gender,
+        _ => PersonalTable.SV.GetFormEntry(Species, Form).Gender,
     };
 }
